Contracts functions have been separated into a separate module for better code readability
